# openapi.yaml
openapi: 3.0.3
info:
  title: Smart Chat RAG Chatbot API
  description: |
    RESTful API for Smart Chat - an AI-powered RAG (Retrieval-Augmented Generation) chatbot platform
    with WordPress integration, multi-bot management, and human handoff capabilities.

    ## Features
    - Multi-tenant bot management
    - RAG-based responses using vector embeddings
    - WordPress plugin integration
    - Human agent handoff
    - Stripe billing integration
    - Rate limiting & abuse detection
    - Streaming responses (SSE)

    ## Authentication
    Most endpoints require authentication via NextAuth session cookies or API tokens.
    WordPress endpoints use custom token format: `userId:botId:secret`

  version: 1.0.0
  contact:
    name: API Support
    email: support@smartchat.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.smartchat.com/api
    description: Production server
...
security:
  - sessionAuth: []
  - tokenAuth: []
paths:
...
  /wordpress/send-message:
    post:
      tags:
        - WordPress
      summary: Send chat message from WordPress
      description: |
        Main endpoint for WordPress plugin to send user messages and receive AI responses.
        Supports both standard JSON responses and Server-Sent Events (SSE) streaming.

        **Rate Limits:** 50 messages per minute per client
        **Abuse Detection:** Automatic blocking for SQL injection, XSS, spam patterns
      security:
        - tokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - message
              properties:
                token:
                  type: string
                  description: WordPress token format (userId:botId:secret)
                  example: "123e4567-e89b-12d3-a456-426614174000:abc-123:secret123"
                message:
                  type: string
                  description: User's message text
                  example: "What are your business hours?"
                sessionId:
                  type: string
                  description: Optional session ID for conversation continuity
                  example: "wp_1234567890_abc123"
                metadata:
                  type: object
                  description: Additional context metadata
                  properties:
                    enableStreaming:
                      type: boolean
                      description: Enable SSE streaming responses
                      default: false
                    language:
                      type: string
                      description: User's language code
                      example: "en"
                    userIp:
                      type: string
                      example: "192.168.1.1"
                    userAgent:
                      type: string
                      example: "Mozilla/5.0..."
                    pageUrl:
                      type: string
                      example: "https://example.com/contact"
      responses:
        '200':
          description: Successful response (non-streaming)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  response:
                    type: string
                    description: Bot's reply message
                    example: "We're open Monday-Friday, 9am-5pm EST"
                  sessionId:
                    type: string
                    example: "wp_1234567890_abc123"
                  conversationId:
                    type: string
                    format: uuid
                    example: "123e4567-e89b-12d3-a456-426614174000"
                  mode:
                    type: string
                    enum: [AI, Human]
                    example: "AI"
                  source:
                    type: string
                    enum: [n8n_rag, openai_direct, openai_fallback]
                    example: "n8n_rag"
                  waitingForAgent:
                    type: boolean
                    description: True if conversation is in human mode
                    example: false
        '200 (streaming)':
          description: SSE streaming response
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream
                example: |
                  data: {"content":"We're","done":false}

                  data: {"content":" open","done":false}

                  data: {"content":"","done":true,"sessionId":"wp_123","conversationId":"abc-123"}
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - abuse detected
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Request blocked"
                  message:
                    type: string
                    example: "Your request was blocked due to suspicious activity."
        '429':
          description: Rate limit exceeded or quota exceeded
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
                example: 50
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: 0
            X-RateLimit-Reset:
              schema:
                type: integer
                example: 1640995200
            Retry-After:
              schema:
                type: integer
                example: 45
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Rate limit exceeded"
                  message:
                    type: string
                    example: "Too many requests. Please try again in 45 seconds."
                  retryAfter:
                    type: integer
                    example: 45
                  upgradeUrl:
                    type: string
                    description: Present if quota exceeded
                    example: "https://app.smartchat.com/manager-dashboard/settings?tab=billing"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: WordPress integration token (userId:botId:secret)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
        message:
          type: string
          example: "Detailed error description"
...